web/requirement.txt
pandas
numpy
streamlit
psycopg2-binary
sqlalchemy